routing and name changes to  pages/components
