+++ client/src/components/AdvocacyCard/BaseAdvocacyCard.js
@@ -5,15 +5,15 @@ import Card from "@material-ui/core/Card";
 import CardActions from "@material-ui/core/CardActions";
 import CardContent from "@material-ui/core/CardContent";
 import Button from "@material-ui/core/Button";
-import GeneralCard from "./GeneralCard";
+import AdvocacyCard from "./AdvocacyCard";
+import LoadingCircle from "./LoadingCircle";
 import thumbtackRed from "../Card/assets/thumbtackred.png";
 import Styled from "styled-components";
-import LoadingCircle from "../LoadingCircle/LoadingCircle";
 import { Link } from "react-router-dom";
 
-const styles = {
+const styles = theme => ({
   card: {
-    minWidth: 275,
+    minWidth: 600,
     width: "100%",
     backgroundColor: "#81c784",
     borderRadius: "10px",
@@ -23,16 +23,18 @@ const styles = {
     margin: "0 2px",
     transform: "scale(0.8)",
   },
+  title: {
+    fontSize: 14,
+  },
   pos: {
     marginBottom: 12,
   },
-};
+});
 
 const linkStyle = {
   textDecoration: "none",
-  backgroundColor: "white",
-  borderRadius: "10px",
-  marginBottom: "10px",
+  backgroundColor:"white",
+  borderRadius: "10px"
 };
 
 const ImageWrapper = Styled.div`
@@ -43,15 +45,15 @@ const ImageWrapper = Styled.div`
   max-width: 100%;
 `;
 
-class BaseGeneralCard extends Component {
+class BaseAdvocacyCard extends Component {
   renderCards = () => {
     const posts = this.props.posts;
     return posts
-      .filter(post => post.category === "General")
+      .filter(post => post.category === "Advocacy")
       .slice(0, 9)
       .map((post, index) => {
         return (
-          <GeneralCard
+          <AdvocacyCard
             key={index}
             id={post.userId}
             name={post.name}
@@ -66,26 +68,24 @@ class BaseGeneralCard extends Component {
 
   render() {
     const { classes } = this.props;
-    // const bull = <span className={classes.bullet}>•</span>;
     return (
       <Card className={classes.card} square>
+        <ImageWrapper>
+          <img
+            className={classes.tack}
+            alt=""
+            src={thumbtackRed}
+            width="50px"
+          />
+        </ImageWrapper>
+        <ImageWrapper>
+          <Link to="/advocacy" style={linkStyle}>
+            <Button color="default" className={classes.button} size="large">
+              {this.props.category}
+            </Button>
+          </Link>
+        </ImageWrapper>
         <CardContent>
-          <ImageWrapper>
-            <img
-              className={classes.tack}
-              alt=""
-              src={thumbtackRed}
-              width="50px"
-            />
-          </ImageWrapper>
-          <ImageWrapper>
-            <Link to="/general" style={linkStyle}>
-              <Button color="default" className={classes.button} size="large">
-                {this.props.category}
-              </Button>
-            </Link>
-          </ImageWrapper>
-
           {this.props.posts.length ? this.renderCards() : <LoadingCircle />}
         </CardContent>
         <CardActions />
@@ -94,8 +94,8 @@ class BaseGeneralCard extends Component {
   }
 }
 
-BaseGeneralCard.propTypes = {
+BaseAdvocacyCard.propTypes = {
   classes: PropTypes.object.isRequired,
 };
 
-export default withStyles(styles)(BaseGeneralCard);
+export default withStyles(styles)(BaseAdvocacyCard);
