+++ client/src/components/AdvocacyPage/PageTabs.js
@@ -5,7 +5,7 @@ import AppBar from "@material-ui/core/AppBar";
 import Tabs from "@material-ui/core/Tabs";
 import Tab from "@material-ui/core/Tab";
 import Typography from "@material-ui/core/Typography";
-import BaseEducationCard from "../EducationCard/BaseEducationCard";
+import BaseAdvocacyCard from "../AdvocacyCard/BaseAdvocacyCard";
 
 function TabContainer(props) {
   return (
@@ -35,10 +35,6 @@ class PageTabs extends React.Component {
     this.setState({ value });
   };
 
-  pageOne = () => {
-    return;
-  };
-
   render() {
     const { classes } = this.props;
     const { value } = this.state;
@@ -64,8 +60,8 @@ class PageTabs extends React.Component {
         </AppBar>
         {value === 0 && (
           <TabContainer>
-            <BaseEducationCard
-              category="Education"
+            <BaseAdvocacyCard
+              category="Advocacy"
               posts={this.props.posts.filter((post, index) => {
                 return index < 9;
               })}
@@ -75,8 +71,8 @@ class PageTabs extends React.Component {
         )}
         {value === 1 && (
           <TabContainer>
-            <BaseEducationCard
-              category="Education"
+            <BaseAdvocacyCard
+              category="Advocacy"
               posts={this.props.posts.filter((post, index) => {
                 return index > 9 && index <= 19;
               })}
@@ -86,8 +82,8 @@ class PageTabs extends React.Component {
         )}
         {value === 2 && (
           <TabContainer>
-            <BaseEducationCard
-              category="Education"
+            <BaseAdvocacyCard
+              category="Advocacy"
               posts={this.props.posts.filter((post, index) => {
                 return index > 19 && index <= 29;
               })}
@@ -97,8 +93,8 @@ class PageTabs extends React.Component {
         )}
         {value === 3 && (
           <TabContainer>
-            <BaseEducationCard
-              category="Education"
+            <BaseAdvocacyCard
+              category="Advocacy"
               posts={this.props.posts.filter((post, index) => {
                 return index > 29 && index <= 39;
               })}
@@ -108,8 +104,8 @@ class PageTabs extends React.Component {
         )}
         {value === 4 && (
           <TabContainer>
-            <BaseEducationCard
-              category="Education"
+            <BaseAdvocacyCard
+              category="Advocacy"
               posts={this.props.posts.filter((post, index) => {
                 return index > 39;
               })}
