+++ client/src/components/CommunityPage/PageTabs.js
@@ -5,7 +5,7 @@ import AppBar from "@material-ui/core/AppBar";
 import Tabs from "@material-ui/core/Tabs";
 import Tab from "@material-ui/core/Tab";
 import Typography from "@material-ui/core/Typography";
-import BaseSkillsCard from "../SkillsCard/BaseSkillsCard";
+import BaseCommunityCard from "../CommunityCard/BaseCommunityCard";
 
 function TabContainer(props) {
   return (
@@ -35,10 +35,6 @@ class PageTabs extends React.Component {
     this.setState({ value });
   };
 
-  pageOne = () => {
-    return;
-  };
-
   render() {
     const { classes } = this.props;
     const { value } = this.state;
@@ -47,7 +43,7 @@ class PageTabs extends React.Component {
       <div className={classes.root}>
         <AppBar position="static" style={{ backgroundColor: "#4caf50" }}>
           <Tabs value={value} onChange={this.handleChange}>
-            <Tab label="Page 1" style={{ color: "#fafafa" }} />
+          <Tab label="Page 1" style={{ color: "#fafafa" }} />
             {this.props.posts.length > 10 ? (
               <Tab label="Page 2" style={{ color: "#fafafa" }} />
             ) : null}
@@ -64,8 +60,8 @@ class PageTabs extends React.Component {
         </AppBar>
         {value === 0 && (
           <TabContainer>
-            <BaseSkillsCard
-              category="Skills/Services"
+            <BaseCommunityCard 
+              category="Community"
               posts={this.props.posts.filter((post, index) => {
                 return index < 9;
               })}
@@ -75,8 +71,8 @@ class PageTabs extends React.Component {
         )}
         {value === 1 && (
           <TabContainer>
-            <BaseSkillsCard
-              category="Skills/Services"
+            <BaseCommunityCard 
+              category="Community"
               posts={this.props.posts.filter((post, index) => {
                 return index > 9 && index <= 19;
               })}
@@ -86,8 +82,8 @@ class PageTabs extends React.Component {
         )}
         {value === 2 && (
           <TabContainer>
-            <BaseSkillsCard
-              category="Skills/Services"
+            <BaseCommunityCard 
+              category="Community"
               posts={this.props.posts.filter((post, index) => {
                 return index > 19 && index <= 29;
               })}
@@ -97,8 +93,8 @@ class PageTabs extends React.Component {
         )}
         {value === 3 && (
           <TabContainer>
-            <BaseSkillsCard
-              category="Skills/Services"
+            <BaseCommunityCard 
+              category="Community"
               posts={this.props.posts.filter((post, index) => {
                 return index > 29 && index <= 39;
               })}
@@ -108,8 +104,8 @@ class PageTabs extends React.Component {
         )}
         {value === 4 && (
           <TabContainer>
-            <BaseSkillsCard
-              category="Skills/Services"
+            <BaseCommunityCard 
+              category="Community"
               posts={this.props.posts.filter((post, index) => {
                 return index > 39;
               })}
