+++ client/src/components/EducationCard/BaseEducationCard.js
@@ -5,8 +5,8 @@ import Card from "@material-ui/core/Card";
 import CardActions from "@material-ui/core/CardActions";
 import CardContent from "@material-ui/core/CardContent";
 import Button from "@material-ui/core/Button";
-import CarPoolCard from "./CarPoolCard";
-import LoadingCircle from "./LoadingCircle";
+import EducationCard from "./EducationCard";
+import LoadingCircle from "../LoadingCircle/LoadingCircle";
 import thumbtackRed from "../Card/assets/thumbtackred.png";
 import Styled from "styled-components";
 import { Link } from "react-router-dom";
@@ -45,15 +45,17 @@ const ImageWrapper = Styled.div`
   max-width: 100%;
 `;
 
-class BaseCarPoolCard extends Component {
+class BaseEducationCard extends Component {
   renderCards = () => {
     const posts = this.props.posts;
     return posts
-      .filter(post => post.category === "Carpool")
+      .filter(post => post.category === "Education")
       .slice(0, 9)
       .map((post, index) => {
+        console.log(post.userId);
+        console.log("posts");
         return (
-          <CarPoolCard
+          <EducationCard
             key={index}
             id={post.userId}
             name={post.name}
@@ -79,7 +81,7 @@ class BaseCarPoolCard extends Component {
           />
         </ImageWrapper>
         <ImageWrapper>
-          <Link to="/carpool" style={linkStyle}>
+          <Link to="/education" style={linkStyle}>
             <Button color="default" className={classes.button} size="large">
               {this.props.category}
             </Button>
@@ -94,8 +96,8 @@ class BaseCarPoolCard extends Component {
   }
 }
 
-BaseCarPoolCard.propTypes = {
+BaseEducationCard.propTypes = {
   classes: PropTypes.object.isRequired,
 };
 
-export default withStyles(styles)(BaseCarPoolCard);
+export default withStyles(styles)(BaseEducationCard);
